Updated issue port data type from uint8_t to uint16_t

+++ src/lib/arch/aarch64/Instruction.cc
@@ -195,7 +195,7 @@ void Instruction::setExecutionInfo(const ExecutionInfo& info) {
   stallCycles_ = info.stallCycles;
   supportedPorts_ = info.ports;
 }
-const std::vector<uint8_t>& Instruction::getSupportedPorts() {
+const std::vector<uint16_t>& Instruction::getSupportedPorts() {
   if (supportedPorts_.size() == 0) {
     exception_ = InstructionException::NoAvailablePort;
     exceptionEncountered_ = true;
